Split prompt and tooltip modules again
There's no reason for these to be combined
