+++ src/prompt/index.js
